+++ changelog.md
@@ -1,4 +1,4 @@
-# Version 2024.4.9 (2024-04-16)
+# Version 2024.4.9 (2024-04-20)
 
 - Decompose combined parameter
 - Return affected entity keys for service calls